Changed airport timezone type to real/float, made correction to save method in AirportAccessor.

+++ src/main/java/seng202/team5/data/AddData.java
@@ -5,7 +5,7 @@ public abstract class AddData {
     public abstract int addAirline(String name, String alias, String iata, String icao, String callsign, String country, String active);
 
     public abstract int addAirport(String name, String city, String country, String iata, String icao, double latitude,
-                                    double longitude, int altitude, int timezone, String dst, String tz);
+                                    double longitude, int altitude, float timezone, String dst, String tz);
 
     public abstract int addFlightEntry(int flightID, String airline, String airport, int altitude, double latitude, double longitude);
